Update on "[BE][DTensor] move torch.distributed._tensor import to torch.distributed.tensor in test files"
cc H-Huang awgu wanchaol fegin fduwjj wz337 wconstab d4l3k

[ghstack-poisoned]
