+++ torch/testing/_internal/distributed/common_state_dict.py
@@ -10,13 +10,13 @@ import torch
 import torch.nn as nn
 from torch.distributed._sharded_tensor import ShardedTensor
 from torch.distributed._state_dict_utils import _gather_state_dict
-from torch.distributed._tensor import DTensor
 from torch.distributed.checkpoint.state_dict import (
     _PG,
     _STATE,
     set_state_dict,
     StateDictOptions,
 )
+from torch.distributed.tensor import DTensor
 
 
 class VerifyStateDictMixin:
